Fixed some exceptions, bugs and added dynamic email to review service for review creation

+++ review-service/src/main/java/com/esiproject2023/reviewservice/service/ReviewService.java
@@ -3,6 +3,7 @@ package com.esiproject2023.reviewservice.service;
 import com.esiproject2023.reviewservice.dto.EmailRequest;
 import com.esiproject2023.reviewservice.dto.MetadataResponse;
 import com.esiproject2023.reviewservice.dto.ReviewDto;
+import com.esiproject2023.reviewservice.dto.UserResponse;
 import com.esiproject2023.reviewservice.model.Review;
 import com.esiproject2023.reviewservice.repository.ReviewRepository;
 import lombok.RequiredArgsConstructor;
@@ -32,7 +33,7 @@ public class ReviewService {
 
     private final KafkaTemplate<String, EmailRequest> kafkaTemplate;
 
-    public ReviewDto createReview(ReviewDto reviewDto) {
+    public String createReview(ReviewDto reviewDto) {
         Review review = Review.builder()
                 .userId(reviewDto.getUserId())
                 .contentId(reviewDto.getContentId())
@@ -41,20 +42,26 @@ public class ReviewService {
                 .build();
 
         MetadataResponse[] response = webClient.build().get().uri("http://metadata-service/metadata/searchByIDs/{ids}", review.getContentId()).retrieve().bodyToMono(MetadataResponse[].class).block();
-        EmailRequest emailRequest = new EmailRequest(
-                "rufatabdullayev029@gmail.com",
-                "Thank you very much for your review!", "You have added review",
-                "\nYou have recently added review/rating to the content " + response[0].getTitle() + "\n\nWe really apprecieate your time spent for making CineMate" +
-                        " better place for entertainment industry!\n\nSincerely,\nTeam CineMate!");
-        Review result = reviewRepository.save(review);
-        kafkaTemplate.send("emailTopic", emailRequest);
+        UserResponse userResponse = webClient.build().get().uri("http://auth-service/auth/{username}", review.getUserId()).retrieve().bodyToMono(UserResponse.class).block();
+        if(response != null && userResponse != null) {
+            EmailRequest emailRequest = new EmailRequest(
+                    userResponse.getEmail(),
+                    "Thank you very much for your review!", "You have added review",
+                    "\nYou have recently added review/rating to the content " + response[0].getTitle() + "\n\nWe really appreciate your time spent for making CineMate" +
+                            " better place for entertainment industry!\n\nSincerely,\nTeam CineMate!");
+            reviewRepository.save(review);
+            log.info("The review with id {} is added", review.getId());
+            kafkaTemplate.send("emailTopic", emailRequest);
+            return "Success";
+        } else return "Fail";
+
 
 //        webClient.build().post().uri("http://email-service/email/send").body(Mono.just(emailRequest), EmailRequest.class).exchangeToMono(emailResponse -> Mono.just(emailResponse.statusCode())).block();
 
 //        webClient.build().post().uri("http://email-service/email/send").body(Mono.just(emailRequest), EmailRequest.class).retrieve().bodyToMono(EmailRequest.class).subscribe();
 
-        log.info("The review with id {} is added", review.getId());
-        return mapToReviewDto(result);
+
+//        return mapToReviewDto(result);
     }
 
     public List<ReviewDto> getAllReviews() {
@@ -81,27 +88,35 @@ public class ReviewService {
         return reviewRepository.findByContentId(contentId);
     }
 
-    public void deleteReview(Long id) {
+    public String deleteReview(Long id) {
         Optional<Review> reviewToBeDeleted = reviewRepository.findById(id);
-        String contentIdForDeletedReview = reviewToBeDeleted.get().getContentId();
+//        reviewToBeDeleted.orElse(Review[].class).getContentId();
         MetadataResponse[] response;
-        if(contentIdForDeletedReview != null) {
-            response = webClient.build().get().uri("http://metadata-service/metadata/searchByIDs/{ids}", contentIdForDeletedReview).retrieve().bodyToMono(MetadataResponse[].class).block();
-            if(response != null) {
-                //           Dynamic email here, once User Auth is done.
-                EmailRequest emailRequest = new EmailRequest(
-                        "muradkhasmammadov@gmail.com",
-                        "We have deleted your review!", "Your review has been deleted",
-                        "We have deleted your review for the content " + "'" + response[0].getTitle() + "'" + ", as it was containing inappropriate content which was against our policy regarding" +
-                                " adding reviews/ratings for the content. We always try to keep our platform safe and friendly for everyone.\n\n" +
-                                "Unfortunately, we will have to ban you from the platform in case of repetition of such case.\n\n" +
-                                "Thank you for your understanding and cooperation.\n\n"+ "Your deleted review for the content was like:\n" + "'" + reviewToBeDeleted.get().getBody() + "'" +
-                                "\n\nSincerely,\nTeam CineMate!");
-                kafkaTemplate.send("emailTopic", emailRequest);
+        if(reviewToBeDeleted.isPresent()) {
+            String contentIdForDeletedReview = reviewToBeDeleted.get().getContentId();
+            if(contentIdForDeletedReview != null) {
+                response = webClient.build().get().uri("http://metadata-service/metadata/searchByIDs/{ids}", contentIdForDeletedReview).retrieve().bodyToMono(MetadataResponse[].class).block();
+                if(response != null) {
+                    //           Dynamic email here, once User Auth is done.
+                    EmailRequest emailRequest = new EmailRequest(
+                            "muradkhasmammadov@gmail.com",
+                            "We have deleted your review!", "Your review has been deleted",
+                            "We have deleted your review for the content " + "'" + response[0].getTitle() + "'" + ", as it was containing inappropriate content which was against our policy regarding" +
+                                    " adding reviews/ratings for the content. We always try to keep our platform safe and friendly for everyone.\n\n" +
+                                    "Unfortunately, we will have to ban you from the platform in case of repetition of such case.\n\n" +
+                                    "Thank you for your understanding and cooperation.\n\n"+ "Your deleted review for the content was like:\n" + "'" + reviewToBeDeleted.get().getBody() + "'" +
+                                    "\n\nSincerely,\nTeam CineMate!");
+                    reviewRepository.deleteById(id);
+                    log.info("Review with id {} has been deleted", id);
+                    kafkaTemplate.send("emailTopic", emailRequest);
+                    return "Success";
+                }
             }
+        } else {
+            log.info("Review not found");
+            return "Fail";
         }
-        reviewRepository.deleteById(id);
-        log.info("Review with id {} has been deleted", id);
+        return "Fail";
     }
 
 }
